+++ engineering-toolkit/eslint-config/eslint.config.js
@@ -1,5 +1,0 @@
-import { concat } from "eslint-flat-config-utils";
-
-import { ecma, style } from "./src/index.js";
-
-export default concat(ecma(), style());