+++ packages/cli/src/domains/generate/async/index.ts
@@ -1,1 +1,0 @@
-export { default as wait } from "./wait";