+++ packages/cli/src/domains/generate/directory/index.ts
@@ -1,3 +1,0 @@
-export { default as existsDirectory } from "./existsDirectory";
-export { default as getDirectory } from "./getDirectory";
-export { default as removeFileOrDirectory } from "./removeFileOrDirectory";