+++ packages/cli/src/domains/generate/directory/removeFileOrDirectory.ts
@@ -1,5 +1,0 @@
-import { rimraf } from "rimraf";
-
-const removeFileOrDirectory = rimraf;
-
-export default removeFileOrDirectory;