+++ packages/cli/src/domains/generate/integration/index.ts
@@ -1,3 +1,0 @@
-export type { default as Integration } from "./Integration";
-export { default as fetchIntegrations } from "./fetchIntegrations";
-export { default as getIntegration } from "./getIntegration";