+++ packages/cli/src/domains/generate/math/identity.ts
@@ -1,8 +1,0 @@
-/**
- * Returns first argument.
- *
- * f(X) = X
- */
-const identity = <T>(value: T): T => value;
-
-export default identity;