+++ packages/middleware/__tests__/integration/bootstrap/serverWithUpload.ts
@@ -1,0 +1,16 @@
+import { apiClientFactory } from "../../../src/apiClientFactory";
+import * as api from "./api";
+
+const onCreate = async (config: Record<string, unknown> = {}) => {
+  return {
+    config,
+    client: null,
+  };
+};
+
+const { createApiClient } = apiClientFactory({
+  onCreate,
+  api,
+});
+
+export { createApiClient };