+++ packages/middleware/src/index.ts
@@ -3,3 +3,4 @@ export * from "./createServer";
 export * from "./apiClientFactory";
 export * from "./terminus";
 export { getLogger } from "./logger";
+export { defaultErrorHandler } from "./errors";